+++ backend/dduishu/src/main/java/com/b208/dduishu/domain/user/entity/UserLevel.java
@@ -19,7 +19,7 @@ public class UserLevel {
     private int exp=0;
 
 
-    public void addExp(int exp){
+    public void addExp(float exp){
         this.exp += exp;
         setUserLevel();
     }
